requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
hydra-core>=1.3.0
wandb>=0.12.0
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.950
onnx>=1.12.0
onnxruntime>=1.12.0
tensorboard>=2.8.0
albumentations>=1.1.0
tqdm>=4.64.0
pyyaml>=6.0
pandas>=1.4.0
scikit-learn>=1.0.0
